Upload ova and qcow improve #1139

+++ src/Controller/OperatingSystemController.php
@@ -409,18 +409,19 @@ class OperatingSystemController extends Controller
                 if ($uploadedFilename && trim($uploadedFilename) !== '') {
                     $localFilePath = $this->getParameter('image_directory') . '/' . $uploadedFilename;
                     $remoteFilePath = '/images/'.$uploadedFilename;
-                    $this->Files2WorkerManager->deleteFileFromAllWorkers($old_filename);
-                    $results=$this->Files2WorkerManager->CopyFileToAllWorkers($localFilePath, $remoteFilePath);
-
-                    $failures = array_filter($results, function($result) {
-                    });
                     
-                    if (!empty($failures)) {
-                        $this->addFlash('warning', 'Image created but some workers failed to send the file.');
-                    } else {
-                        unlink($this->getParameter('image_directory') . '/' . $old_filename);
-                        $this->addFlash('success', 'Image created and file copied to all workers successfully.');
-                    }
+                    if ($this->Files2WorkerManager->AvailableWorkerExist()) {
+                        $this->Files2WorkerManager->deleteFileFromAllWorkers($old_filename);
+                        $results=$this->Files2WorkerManager->CopyFileToAllWorkers($localFilePath, $remoteFilePath);
+                        $failures = array_filter($results, function($result) { 
+                        });
+                        if (!empty($failures)) {
+                            $this->addFlash('warning', 'Image created but some workers failed to send the file.');
+                        } else {
+                            unlink($this->getParameter('image_directory') . '/' . $old_filename);
+                            $this->addFlash('success', 'Image created and file copied to all workers successfully.');
+                        }
+                    }                   
                 }
             } else {
                 // Hyperviseur non reconnu - nettoyer les champs image
@@ -632,86 +633,113 @@ class OperatingSystemController extends Controller
      */
     #[Route('/api/images/upload', name: 'app_images_upload', methods: ['POST'])]
     public function upload(Request $request, SluggerInterface $slugger): Response
-    {
-        $this->logger->info('Uploading image file requested by user '.$this->getUser()->getName());
-        
-        $file = $request->files->get('file');
-        
-        if ($file && $file->isValid()) {
-            $this->logger->debug('[OperatingSystemController:upload]::The file to upload will be '.$file.' of size '.$file->getSize());
+{
+    $this->logger->info('Uploading image file requested by user '.$this->getUser()->getName());
+    
+    $file = $request->files->get('file');
+    
+    if ($file && $file->isValid()) {
+        $this->logger->debug('[OperatingSystemController:upload]::The file to upload will be '.$file.' of size '.$file->getSize());
 
-            $maxUploadSize = $this->convertPHPSizeToBytes(min(ini_get('upload_max_filesize'),ini_get('post_max_size')));
+        $maxUploadSize = $this->convertPHPSizeToBytes(min(ini_get('upload_max_filesize'),ini_get('post_max_size')));
 
-            if ($file->getSize() > $maxUploadSize) {
-                return $this->json(['error' => 'File too large'], 413);
-            }
-
-            $originalFilename = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-            $safeFilename = $slugger->slug($originalFilename);
-            $newFilename = $safeFilename . '-' . uniqid() . '.' . $file->guessExtension();
-            $file_extension=$file->guessExtension();
-            
-            if ($file_extension === "tar" || $file_extension === "img" || $file_extension === "qcow2") {
-                $filesystem = new Filesystem();
-                if ($file->guessExtension()==="tar") {
-                    $this->logger->debug('[OperatingSystemController:upload]::Detect the '.$file.' is ova');
-                    $newFilename = basename($newFilename,".tar").".qcow2";
-                    $temp_path=sys_get_temp_dir()."/Remotelabz-image-convert-".uniqid();
-                    $filesystem->mkdir($temp_path);
-                    
-                    // The file stays in the temp directory
-                    if ($this->convertOva2qcow2($temp_path,$file, $newFilename))
-                        $this->logger->debug('[OperatingSystemController:upload]::Convert '.$temp_path."/".basename($file).' file to '.$temp_path."/".$newFilename.' done.');
-                    else
-                        throw new \Exception('Error during OVA to QCOW2 conversion');
-                }
-                try {
-                    $uploadDir = $this->getParameter('image_directory');
-                    $this->logger->debug('[OperatingSystemController:upload]::Try to move '.$temp_path."/".$newFilename.' file to '.$uploadDir);        
-                    $file->move($temp_path."/".$newFilename,$uploadDir);
-            
-                    $newFile = $uploadDir . '/' . $newFilename;
-                    $fileSize = file_exists($newFile) ? filesize($newFile) : null;
-
-                    $this->logger->debug('[OperatingSystemController:upload]::Move '.$temp_path."/".$newFilename.' file to '.$uploadDir.' done');        
-                    $this->logger->info('Move '.$temp_path."/".$newFilename.' file to '.$uploadDir.' done');        
-                    if ($filesystem->exists($temp_path)) {
-                        $filesystem->remove($temp_path);
-                        $this->logger->debug('[OperatingSystemController:upload]::Temporary directory cleaned');
-                    }
-                    return $this->json([
-                        'success' => true, 
-                        'filename' => $newFilename,
-                        'originalName' => $file->getClientOriginalName(),
-                        'size' => $fileSize
-                    ]);
-                }
-                catch (FileException $e) {
-                    $this->logger->error('[OperatingSystemController:upload]::FileException occurs: ' . $e->getMessage());
-                    
-                    if (str_contains($e->getMessage, "No space left"))
-                        return $this->json(['error' => 'No space left on the device'], 500);        
-                    else
-                        return $this->json(['error' => 'Upload failed'], 500);
-                }
-                catch (\Exception $e) {
-                    $this->logger->error('[OperatingSystemController:upload]::Error during file upload: ' . $e->getMessage());
-                    return $this->json(['error' => 'Upload failed due to an unexpected error'], 500);
-                }
-                finally {
-                    // --- Nettoyage (toujours exécuté) ---
-                    if ($filesystem->exists($temp_path)) {
-                        $filesystem->remove($temp_path);
-                        $this->logger->debug('[OperatingSystemController:upload]::Temporary directory cleaned');
-                    }
-                }
-            } else {
-                    $this->logger->debug('[OperatingSystemController:upload]::The upload file is not an ova but a '.$file_extension);
-                    return $this->json(['error' => 'OVA files are not supported for upload. Please use IMG or QCOW2 files.'], 400);
-            }
+        if ($file->getSize() > $maxUploadSize) {
+            return $this->json(['error' => 'File too large'], 413);
         }
-        else return $this->json(['error' => 'No file uploaded'], 400);
+
+        $originalFilename = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
+        $safeFilename = $slugger->slug($originalFilename);
+        $newFilename = $safeFilename . '-' . uniqid() . '.' . $file->guessExtension();
+        $file_extension = $file->guessExtension();
+        $this->logger->debug('[OperatingSystemController:upload]::The '.$file.' is a '.$file_extension.' file');
+        
+        if ($file_extension === "tar" || $file_extension === "img" || $file_extension === "qcow2") {
+            $filesystem = new Filesystem();
+            $temp_path = sys_get_temp_dir()."/Remotelabz-image-convert-".uniqid();
+            $filesystem->mkdir($temp_path);
+            
+            try {
+                $uploadDir = $this->getParameter('image_directory');
+                
+                if ($file_extension === "tar") {
+                    // Conversion OVA vers QCOW2
+                    $newFilename = basename($newFilename, ".tar") . ".qcow2";
+                    $converted_file=$this->convertOva2qcow2($temp_path, $file, $newFilename);
+
+                    if ($converted_file) {
+                        $this->logger->debug('[OperatingSystemController:upload]::Convert OVA file to '.$converted_file.' done.');
+                    } else {
+                        throw new \Exception('Error during OVA to QCOW2 conversion');
+                    }
+                    
+                    // Déplacement du fichier converti vers le répertoire final
+                    
+                    $destinationFile = $uploadDir . "/" . $newFilename;
+                    
+                    $this->logger->debug('[OperatingSystemController:upload]::Try to move '.$converted_file.' file to '.$destinationFile);
+
+                    if (!file_exists($converted_file)) {
+                        throw new \Exception('Converted file not found: ' . $converted_file);
+                    }
+
+                    if (!rename($converted_file, $destinationFile)) {
+                        // Si rename échoue (par exemple entre partitions différentes), utiliser copy + unlink
+                        if (!copy($converted_file, $destinationFile)) {
+                            throw new \Exception('Failed to copy file to destination');
+                        }
+                        unlink($converted_file);
+                    }
+                    
+                } else {
+                    // Pour IMG et QCOW2, déplacement direct
+                    $this->logger->debug('[OperatingSystemController:upload]::Try to move file to '.$uploadDir);
+                    $file->move($uploadDir, $newFilename);
+                    $destinationFile = $uploadDir . "/" . $newFilename;
+                }
+                
+                $fileSize = file_exists($destinationFile) ? filesize($destinationFile) : null;
+                
+                $this->logger->debug('[OperatingSystemController:upload]::Move file to '.$uploadDir.' done');
+                $this->logger->info('File successfully uploaded to '.$uploadDir);
+                
+                return $this->json([
+                    'success' => true, 
+                    'filename' => $newFilename,
+                    'originalName' => $file->getClientOriginalName(),
+                    'size' => $fileSize
+                ]);
+            }
+            catch (FileException $e) {
+                $this->logger->error('[OperatingSystemController:upload]::FileException occurs: ' . $e->getMessage());
+                
+                if (str_contains($e->getMessage(), "No space left"))
+                    return $this->json(['error' => 'No space left on the device'], 500);
+                else
+                    return $this->json(['error' => 'Upload failed'], 500);
+            }
+            catch (\Exception $e) {
+                $this->logger->error('[OperatingSystemController:upload]::Error during file upload: ' . $e->getMessage());
+                if (str_contains($e->getMessage(), "No space left"))
+                    return $this->json(['error' => 'No space left on the device'], 500);
+                else
+                    return $this->json(['error' => 'Upload failed due to an unexpected error'], 500);
+            }
+            finally {
+                // Nettoyage du répertoire temporaire
+                if ($filesystem->exists($temp_path)) {
+                    $filesystem->remove($temp_path);
+                    $this->logger->debug('[OperatingSystemController:upload]::Temporary directory '.$temp_path.' cleaned');
+                }
+            }
+        } else {
+            $this->logger->debug('[OperatingSystemController:upload]::The upload file is not supported: '.$file_extension);
+            return $this->json(['error' => 'Only TAR (OVA), IMG or QCOW2 files are supported for upload.'], 400);
+        }
     }
+    else {
+        return $this->json(['error' => 'No file uploaded'], 400);
+    }
+}
 
     #[Route('/api/images/delete-temp-file', name: 'app_images_delete_temp_file', methods: ['POST','DELETE'])]
     public function deleteTempFile(Request $request): Response
@@ -933,7 +961,7 @@ class OperatingSystemController extends Controller
         return $bytes;
     }
 
-    private function convertOva2qcow2(string $tempDir,string $filename, string $newfilename): bool
+    private function convertOva2qcow2(string $tempDir,string $filename, string $newfilename): string
     {
         $filesystem = new Filesystem();
         
@@ -950,9 +978,11 @@ class OperatingSystemController extends Controller
                 $this->logger->debug('[OperatingSystemController:convertOva2qcow2]::Tar executed');
                 
                 // Convert all vmdk in one qcow2
-                $this->ovaManager->processVmdkFiles($tempDir);
+                $newfilename=$this->ovaManager->processVmdkFiles($tempDir);
+                //newfilename with path
+                $this->logger->debug('[OperatingSystemController:convertOva2qcow2]::tempDir:'.$tempDir.' filename:'.$filename.' newfilename:'.$newfilename);
                 $this->logger->debug('[OperatingSystemController:convertOva2qcow2]::VMDK processed');                
-                return true;
+                return $newfilename;
                 
             } catch (ProcessFailedException $exception) {
                 $this->logger->error('Error during OVA to QCOW2 conversion: ' . $exception->getMessage());
